蓝盾流水线自动打准备发布灰度tag - gray/v1.4.1.3-2021-03-25

+++ src/backend/ci/ext/tencent/external/api-external/src/main/kotlin/com/tencent/devops/external/api/ServiceGithubResource.kt
@@ -109,7 +109,7 @@ interface ServiceGithubResource {
         projectName: String,
         @ApiParam("checkRunId", required = true)
         @QueryParam("checkRunId")
-        checkRunId: Int,
+        checkRunId: Long,
         checkRuns: GithubCheckRuns
     ): Result<Boolean>
